+++ tftrt/benchmarking-python/benchmark_info.py
@@ -7,6 +7,11 @@ import os
 import subprocess
 import shlex
 
+# The `__version__` number shall be updated everytime core benchmarking files
+# are updated.
+# Please update CHANGELOG.md with a description of what this version changed.
+__version__ = "1.1.0"
+
 
 def get_commit_id():