+++ tftrt/benchmarking-python/nvidia_examples/electra_tf2/infer.py
@@ -45,6 +45,10 @@ sys.path.insert(0, benchmark_base_dir)
 from benchmark_args import BaseCommandLineAPI
 from benchmark_runner import BaseBenchmarkRunner
 
+from benchmark_logger import logging
+
+SAMPLES_IN_DATASET = 10950
+
 
 class CommandLineAPI(BaseCommandLineAPI):
 
@@ -153,6 +157,15 @@ class CommandLineAPI(BaseCommandLineAPI):
             "models, False for cased models."
         )
 
+    def _post_process_args(self, args):
+        args = super(CommandLineAPI, self)._post_process_args(args)
+        args.num_warmup_iterations = min(
+            int(SAMPLES_IN_DATASET / (args.batch_size) / 2),
+            args.num_warmup_iterations
+        )
+
+        return args
+
 
 # %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%% #
 # %%%%%%%%%%%%%%%%% IMPLEMENT MODEL-SPECIFIC FUNCTIONS HERE %%%%%%%%%%%%%%%%%% #
@@ -178,40 +191,44 @@ class BenchmarkRunner(BaseBenchmarkRunner):
 
         def get_dataset_from_features(features, batch_size):
 
-            all_unique_ids = tf.convert_to_tensor([
-                f.unique_id for f in features
-            ],
-                                                  dtype=tf.int64)
-            all_input_ids = tf.convert_to_tensor([
-                f.input_ids for f in features
-            ],
-                                                 dtype=tf.int64)
-            all_input_mask = tf.convert_to_tensor([
-                f.attention_mask for f in features
-            ],
-                                                  dtype=tf.int64)
-            all_segment_ids = tf.convert_to_tensor([
-                f.token_type_ids for f in features
-            ],
-                                                   dtype=tf.int64)
-            all_start_pos = tf.convert_to_tensor([
-                f.start_position for f in features
-            ],
-                                                 dtype=tf.int64)
-            all_end_pos = tf.convert_to_tensor([
-                f.end_position for f in features
-            ],
-                                               dtype=tf.int64)
-            all_cls_index = tf.convert_to_tensor([
-                f.cls_index for f in features
-            ],
-                                                 dtype=tf.int64)
-            all_p_mask = tf.convert_to_tensor([f.p_mask for f in features],
-                                              dtype=tf.float32)
-            all_is_impossible = tf.convert_to_tensor([
-                f.is_impossible for f in features
-            ],
-                                                     dtype=tf.float32)
+            # yapf: disable
+            all_unique_ids = tf.convert_to_tensor(
+                [f.unique_id for f in features],
+                dtype=tf.int64
+            )
+            all_input_ids = tf.convert_to_tensor(
+                [f.input_ids for f in features],
+                dtype=tf.int64
+            )
+            all_input_mask = tf.convert_to_tensor(
+                [f.attention_mask for f in features],
+                dtype=tf.int64
+            )
+            all_segment_ids = tf.convert_to_tensor(
+                [f.token_type_ids for f in features],
+                dtype=tf.int64
+            )
+            all_start_pos = tf.convert_to_tensor(
+                [f.start_position for f in features],
+                dtype=tf.int64
+            )
+            all_end_pos = tf.convert_to_tensor(
+                [f.end_position for f in features],
+                dtype=tf.int64
+            )
+            all_cls_index = tf.convert_to_tensor(
+                [f.cls_index for f in features],
+                dtype=tf.int64
+            )
+            all_p_mask = tf.convert_to_tensor(
+                [f.p_mask for f in features],
+                dtype=tf.float32
+            )
+            all_is_impossible = tf.convert_to_tensor(
+                [f.is_impossible for f in features],
+                dtype=tf.float32
+            )
+            # yapf: enable
 
             dataset = tf.data.Dataset.from_tensor_slices((
                 all_unique_ids, all_input_ids, all_input_mask, all_segment_ids,
@@ -234,7 +251,6 @@ class BenchmarkRunner(BaseBenchmarkRunner):
 
         dev_examples = processor.get_dev_examples(self._args.data_dir)
 
-        print("***** Loading features *****")
         # Load cached features
         squad_version = '2.0' if self._args.version_2_with_negative else '1.1'
         cache_dir = self._args.data_dir
@@ -263,7 +279,7 @@ class BenchmarkRunner(BaseBenchmarkRunner):
                 )
             )
 
-            print(f"**** Building Cache Files: {cached_dev_features_file} ****")
+            logging.debug(f"Building Cache Files: {cached_dev_features_file}")
             with open(cached_dev_features_file, "wb") as writer:
                 pickle.dump(dev_features, writer)
 
@@ -384,16 +400,15 @@ class BenchmarkRunner(BaseBenchmarkRunner):
             f"{os.path.join(self._args.data_dir, dev_file)} "
             f"{output_prediction_file}"
         )
-        if self._args.debug:
-            print(f"\nExecuting: `{command_str}`\n")
+
+        logging.debug(f"\nExecuting: `{command_str}`\n")
 
         eval_out = subprocess.check_output(shlex.split(command_str))
 
         # scores: {'exact_match': 87.06717123935667, 'f1': 92.78048326711645}
         scores = json.loads(eval_out.decode("UTF-8").strip())
 
-        if self._args.debug:
-            print("scores:", scores)
+        logging.debug("scores:", scores)
 
         metric_units = "f1"