+++ tftrt/benchmarking-python/tf_hub/movinet/models/movinet_a3/run_inference.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+set -x
+
+BASE_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )/../.."
+
+bash ${BASE_DIR}/base_run_inference.sh --model_name="a3" --num_frames=12 --input_size=256 ${@}