+++ tftrt/benchmarking-python/tf_hub/movinet/models/movinet_a4/run_inference.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+set -x
+
+BASE_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )/../.."
+
+bash ${BASE_DIR}/base_run_inference.sh --model_name="a4" --num_frames=8 --input_size=290 ${@}